% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_data_structure.R
\name{csm_create_data_structure}
\alias{csm_create_data_structure}
\title{Create a cropping systems model (CSM) data structure}
\usage{
csm_create_data_structure(name, definition, variables, n_dim = 1)
}
\arguments{
\item{name}{a length-one character vector name of a variable}

\item{definition}{the definition of the data structure}

\item{variables}{a list of the CSM variables (as defined with
[csmbuilder::csm_create_variable()]) that are contained within the data
 structure}

\item{n_dim}{the number of dimensions in the data structure; default value
of 1 (i.e. each variable is expected to be a scalar value)}
}
\value{
a list of `csm_data_structure` objects
}
\description{
Create a cropping systems model (CSM) data structure
}
\examples{

# Create variables:
wth_variables <- c(
  csm_create_variable("time",
                      "time of observation",
                      "days"),
  csm_create_variable("Tair",
                      "air temperature",
                      "Celsius"),
  csm_create_variable("SRAD",
                      "solar radiation",
                      "MJ/m2/d"))

# Create weather data structure:
weather <- csm_create_data_structure("weather",
                                     "weather data",
                                     wth_variables,
                                     n_dim = 2)

}
