% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_dydt.R
\name{csm_create_dydt}
\alias{csm_create_dydt}
\title{Create a function for rate of change of
Cropping System Model (CSM) state variables}
\usage{
csm_create_dydt(
  name,
  state,
  ...,
  arg_names = NULL,
  fill_in_names = FALSE,
  output_type = c("Rfunction", "Rcode", "deSolve"),
  comment_char = "#",
  line_end = "",
  v_types = c(scalar = "", vector = "")
)
}
\arguments{
\item{name}{name of the resulting function}

\item{state}{a list vector containing CSM state
variables defined using \code{\link[=csm_create_state]{csm_create_state()}}
in the intended order}

\item{...}{optional arguments of list vectors containing
CSM parameters defined using \code{\link[=csm_create_parameter]{csm_create_parameter()}}
or CSM variables defined using \code{\link[=csm_create_variable]{csm_create_variable()}}}

\item{arg_names}{an optional named character vector whose elements
correspond to the argument name in the function to be
generated and whose names correspond to either \code{state}
or the name of the argument in \code{...}}

\item{fill_in_names}{a TRUE/FALSE value indicating
whether to fill in default argument names (e.g. arg_2)
for variables in \code{...} that are not specified in
\code{arg_names}}

\item{output_type}{a character value indicating the
type of output to produce}

\item{comment_char}{a character value indicating
which character to use as a prefix to comments
in the generated code}

\item{line_end}{a character value indicating which
character(s) to indicate the end of a line of
source code}

\item{v_types}{a named character vector indicating
the names of the variable types for scalars and
vectors in the target output language}
}
\description{
Create a function for rate of change of
Cropping System Model (CSM) state variables
}
