% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_model.R
\name{csm_create_model}
\alias{csm_create_model}
\title{Create a model definition for a Cropping System Model (CSM)}
\usage{
csm_create_model(state, ..., name = "model")
}
\arguments{
\item{state}{a list vector containing CSM state
variables defined using \code{\link[=csm_create_state]{csm_create_state()}}
in the intended order}

\item{...}{optional arguments of list vectors containing
CSM parameters defined using \code{\link[=csm_create_parameter]{csm_create_parameter()}}
or CSM variables defined using \code{\link[=csm_create_variable]{csm_create_variable()}}}

\item{name}{a character string containing a name for the models}
}
\value{
a list which defines all components of a model including state variables,
input variables, parameters, transformed variables and data structures.
}
\description{
Create a model definition for a Cropping System Model (CSM)
}
\examples{

# Define state variables
lv_state <- csm_create_state(
  c("x", "y"),
  definition = c("prey", "predator"),
  units = c("rabbits per square km", "foxes per square km"),
  expression(~alpha*x-beta*x*y, ~delta*x*y-gamma*y))

# Define parameters
lv_parameters <- csm_create_parameter(
  c("alpha", "beta", "gamma", "delta"),
  definition = c("maximum prey per capita growth rate",
                 "effect of predator population on prey death rate",
                 "predator per capita death rate",
                 "effect of prey population on predator growth rate"),
  units = c("rabbits per rabbit", "per fox",
            "foxes per fox", "foxes per rabbit"))

# Define model
lotka_volterra_model <-
  csm_create_model(
    state = lv_state,
    parms = lv_parameters)

}
