% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_transform.R
\name{csm_create_transform}
\alias{csm_create_transform}
\title{Create a cropping systems model (CSM) state variable}
\usage{
csm_create_transform(name, definition, units, equation)
}
\arguments{
\item{name}{a length-one character vector name of a variable}

\item{units}{a length-one character vector of the units of
the CSM variable}

\item{equation}{an R expression with the equation
for the value of the transformed CSM state variable}
}
\description{
Create a cropping systems model (CSM) state variable
}
