% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_variable.R
\name{csm_create_variable}
\alias{csm_create_variable}
\title{Create a cropping systems model (CSM) variable}
\usage{
csm_create_variable(name, definition, units)
}
\arguments{
\item{name}{a length-one character vector name of a variable}

\item{units}{a length-one character vector of the units of
the CSM variable}

\item{definintion}{a length-one character vector that defines
the CSM variable}
}
\description{
Create a cropping systems model (CSM) variable
}
\examples{

Tair <- csm_create_variable(name = "Tair",
                            definition = "air temperature",
                            units = "Celsius")

}
