% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_dependency_graph.R
\name{csm_dependency_graph}
\alias{csm_dependency_graph}
\title{Create a dependency graph for a Cropping System Model (CSM)}
\usage{
csm_dependency_graph(model)
}
\arguments{
\item{model}{a model as defined using \code{\link[=csm_create_model]{csm_create_model()}}}
}
\description{
Create a dependency graph for a Cropping System Model (CSM)
}
