% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csm_get_at_t}
\alias{csm_get_at_t}
\title{Linearly interpolate a time-varying variable at specific time point}
\usage{
csm_get_at_t(
  x,
  t_ind,
  t,
  method = "linear",
  search = c("interpolation", "bisection", "bruteforce", "i=t+1")
)
}
\arguments{
\item{x}{a vector of a time-varying values for which to interpolate}

\item{t_ind}{a vector of times corresponding to the values in x}

\item{t}{a single time point at which to return a value}

\item{method}{a string indicating what method to use for
interpolation. One of: "linear"}

\item{search}{a string indicating what method to use for
finding the correct indices within t_ind. One of: "bisection",
"interpolation"}
}
\value{
the numeric value of the time-varying variable interpolated at time t
}
\description{
Linearly interpolate a time-varying variable at specific time point
}
