% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csm_get_t_indices}
\alias{csm_get_t_indices}
\title{Get time indices bracketing a specific time point}
\usage{
csm_get_t_indices(
  t_ind,
  t,
  method = "linear",
  search = c("interpolation", "bisection", "bruteforce", "i=t+1")
)
}
\arguments{
\item{t_ind}{a vector of times}

\item{t}{a single time point at which to find the bracketing time
indices}

\item{method}{a string indicating what method to use for
interpolation. One of: "linear"}

\item{search}{a string indicating what method to use for
finding the correct indices within t_ind. One of: "bisection",
"interpolation", "bruteforce", "i=t+1"}
}
\description{
Get time indices bracketing a specific time point
}
