% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csm_hill_down_reg}
\alias{csm_hill_down_reg}
\title{Hill equation for down-regulation}
\usage{
csm_hill_down_reg(L, K, n)
}
\arguments{
\item{L}{a numeric value providing the ligand concentration}

\item{K}{a numeric value providing the ligand concentration at half occupation}

\item{n}{a numeric value providing the Hill coefficient}
}
\value{
a numeric value between 0 and 1
}
\description{
Hill equation for down-regulation
}
