% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csm_mod_arr}
\alias{csm_mod_arr}
\title{Modified Arrhenius function}
\usage{
csm_mod_arr(Tt, ko, H, E, To)
}
\arguments{
\item{Tt}{temperature in Celsius}

\item{ko}{reaction rate at the optimum temperature (To)}

\item{H}{deactivation energy parameter}

\item{E}{activation energy parameter}

\item{To}{optimum temperature in Celsius}
}
\value{
a numeric value of the reaction rate at temperature Tt
}
\description{
Modified Arrhenius function
}
