% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_time_vector.R
\name{csm_time_vector}
\alias{csm_time_vector}
\title{Generate vector of times for simulation}
\usage{
csm_time_vector(t_init, t_max, dt = 1)
}
\arguments{
\item{t_init}{a numerical value providing the initial time to
use for simulation}

\item{t_max}{a numerical value providing the final time to
use for simulation}

\item{dt}{a numerical value providing the size of the time step to
use for simulation}
}
\value{
a numeric vector
}
\description{
Generate vector of times for simulation
}
\examples{

csm_time_vector(0, 100, dt = 0.01)


}
