% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_at_t_linear}
\alias{get_at_t_linear}
\title{Linearly interpolate a time-varying variable at specific time point}
\usage{
get_at_t_linear(x, t_ind, t)
}
\arguments{
\item{x}{a vector of a time-varying values for which to interpolate}

\item{t_ind}{a vector of times corresponding to the values in x}

\item{t}{a single time point at which to}
}
\description{
Linearly interpolate a time-varying variable at specific time point
}
