% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_input}
\alias{is_input}
\title{Determine if object is a model input}
\usage{
is_input(x)
}
\arguments{
\item{x}{any R object}
}
\value{
a logical value indicating whether or not x is an input variable
}
\description{
Determine if object is a model input
}
