% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_parameter}
\alias{is_parameter}
\title{Determine if object is a model parameter}
\usage{
is_parameter(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is a model parameter
}
