% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_state_variable}
\alias{is_state_variable}
\title{Determine if object is state variable}
\usage{
is_state_variable(x)
}
\arguments{
\item{x}{any R object}
}
\value{
a logical value indicating whether or not x is of class csm_state
}
\description{
Determine if object is state variable
}
