% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_transform}
\alias{is_transform}
\title{Determine if object is a transformed variable}
\usage{
is_transform(x)
}
\arguments{
\item{x}{any R object}
}
\value{
a logical value indicating whether or not x is of class csm_transform
}
\description{
Determine if object is a transformed variable
}
