% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_data_structure.R
\name{csm_create_data_structure}
\alias{csm_create_data_structure}
\title{Create a cropping systems model (CSM) data structure}
\usage{
csm_create_data_structure(name, definition, variables)
}
\arguments{
\item{name}{a length-one character vector name of a variable}

\item{definition}{the definition of the data structure}

\item{variables}{a list of the CSM variables (as defined with
[csmdeveloper::csm_create_variable()]) that are contained within the data
 structure}
}
\description{
Create a cropping systems model (CSM) data structure
}
