% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_dydt.R
\name{csm_create_dydt}
\alias{csm_create_dydt}
\title{Create a function for rate of change of
Cropping System Model (CSM) state variables}
\usage{
csm_create_dydt(
  name,
  state,
  arg_names = c(state = "y"),
  ...,
  output_type = c("Rfunction", "Rcode", "deSolve"),
  comment_char = "#",
  line_end = "",
  v_types = c(scalar = "", vector = "")
)
}
\arguments{
\item{name}{name of the resulting function}

\item{arg_names}{a named character vector whose elements
correspond to the argument name in the function to be
generated and whose names correspond to either \code{states}
or the name of the argument in \code{...}}

\item{...}{optional arguments of list vectors containing
CSM parameters defined using \code{\link[=csm_create_parameter]{csm_create_parameter()}}
or CSM variables defined using \code{\link[=csm_create_variable]{csm_create_variable()}}}

\item{states}{a list vector containing CSM state
variables defined using \code{\link[=csm_create_state]{csm_create_state()}}
in the intended order}
}
\description{
Create a function for rate of change of
Cropping System Model (CSM) state variables
}
