% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_model.R
\name{csm_create_model}
\alias{csm_create_model}
\title{Create a model definition for a Cropping System Model (CSM)}
\usage{
csm_create_model(state, ..., name = "model")
}
\arguments{
\item{state}{a list vector containing CSM state
variables defined using \code{\link[=csm_create_state]{csm_create_state()}}
in the intended order}

\item{...}{optional arguments of list vectors containing
CSM parameters defined using \code{\link[=csm_create_parameter]{csm_create_parameter()}}
or CSM variables defined using \code{\link[=csm_create_variable]{csm_create_variable()}}}
}
\description{
Create a model definition for a Cropping System Model (CSM)
}
