% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_parameter.R
\name{csm_create_parameter}
\alias{csm_create_parameter}
\title{Create a cropping systems model (CSM) parameter variable}
\usage{
csm_create_parameter(
  name,
  definition,
  units,
  lower_bound = NULL,
  upper_bound = NULL
)
}
\arguments{
\item{name}{a length-one character vector name of a variable}

\item{units}{a length-one character vector of the units of
the CSM variable}

\item{lower_bound}{a numerical value providing the lower
bound for the parameter}

\item{upper_bound}{a numerical value providing the upper
bound for the parameter}
}
\description{
Create a cropping systems model (CSM) parameter variable
}
