% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_create_state.R
\name{csm_create_state}
\alias{csm_create_state}
\title{Create a cropping systems model (CSM) state variable}
\usage{
csm_create_state(name, definition, units, equation)
}
\arguments{
\item{name}{a length-one character vector name of a variabl}

\item{units}{a length-one character vector of the units of
the CSM variable}

\item{equation}{an R expression with the equation
for the rate of change of the CSM state variable}
}
\description{
Create a cropping systems model (CSM) state variable
}
