% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_render_model.R
\name{csm_render_model}
\alias{csm_render_model}
\title{Create a model definition for a Cropping System Model (CSM)}
\usage{
csm_render_model(model)
}
\arguments{
\item{model}{a list vector containing a CSM as created by
\code{\link[=csm_create_model]{csm_create_model()}}}
}
\description{
Create a model definition for a Cropping System Model (CSM)
}
