% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_render_model.R
\name{csm_render_model}
\alias{csm_render_model}
\title{Render a defined Cropping System Model (CSM)}
\usage{
csm_render_model(
  model,
  name = "dy_dt",
  output_type = c("function", "code"),
  language = c("R", "Rcpp"),
  arg_alias = NULL,
  insert_functions = NULL
)
}
\arguments{
\item{model}{a list vector containing a CSM as created by
\code{\link[=csm_create_model]{csm_create_model()}}}

\item{name}{name of the resulting function}

\item{output_type}{a character value indicating the type of output to produce;
one of: "function" (a callable function) or "code" (computer code for the
model)}

\item{language}{a character value indicating which programming language into
which to render the model}

\item{arg_alias}{an optional named character vector whose names indicate
variables for which to use an alias within the generated function and whose
elements provide the corresponding alias}

\item{insert_functions}{an optional list of functions to add to rendered code}
}
\value{
Either an R function object (if \code{output_type}="function") or a character
vector of model code (if \code{output_type}="code") in the programming language
specified by \code{language}.
}
\description{
Render a defined Cropping System Model (CSM)
}
\examples{

# Define state variables
lv_state <- csm_create_state(
  c("x", "y"),
  definition = c("prey", "predator"),
  units = c("rabbits per square km", "foxes per square km"),
  expression(~alpha*x-beta*x*y, ~delta*x*y-gamma*y))

# Define parameters
lv_parameters <- csm_create_parameter(
  c("alpha", "beta", "gamma", "delta"),
  definition = c("maximum prey per capita growth rate",
                 "effect of predator population on prey death rate",
                 "predator per capita death rate",
                 "effect of prey population on predator growth rate"),
  units = c("rabbits per rabbit", "per fox",
            "foxes per fox", "foxes per rabbit"))

# Define model
lotka_volterra_model <-
  csm_create_model(
    state = lv_state,
    parms = lv_parameters)

# Render model into raw R code
lotka_volterra_code <-
  csm_render_model(lotka_volterra_model,
                   output_type = "code",
                   language = "R")

# Render model into a callable R function
lotka_volterra_fun <-
  csm_render_model(lotka_volterra_model,
                   output_type = "function",
                   language = "R")

}
