% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_data_structure}
\alias{is_data_structure}
\title{Determine if object is a data structure}
\usage{
is_data_structure(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Determine if object is a data structure
}
